Update Game and Board classes for One Turn. Update main file for testing
Modified output for game menu when commencing one player turn. Loop seems to be working whenever the player is choosing an action. The game still does not work completely as intended. When rolling certain pairs, the game will call a "bust" even if one pair has a tower already within it that is capable of moving. I believe the logic lies somewhere within the Board::move function.