Add CList class files. Update main file for testing
Added CList header and cpp files. So far, the list seems to be able to properly keep track of the head, tail, current, and previous-current Cells using four pointers and the desired functions. Tests have been implemented within the main file although the output for the .txt file can probably be improved. The CList is also using the Player class by default and it may need some updates for using templates and dynamically adjusting to other types.